+++ backend/substrapp/views/algo.py
@@ -3,6 +3,7 @@ from django.conf import settings
 from django.db import models
 from django.urls import reverse
 from django_filters.rest_framework import BaseInFilter
+from django_filters.rest_framework import CharFilter
 from django_filters.rest_framework import DateTimeFromToRangeFilter
 from django_filters.rest_framework import DjangoFilterBackend
 from django_filters.rest_framework import FilterSet
@@ -26,6 +27,7 @@ from substrapp.views.filters_utils import CustomSearchFilter
 from substrapp.views.filters_utils import ProcessPermissionFilter
 from substrapp.views.utils import CP_BASENAME_PREFIX
 from substrapp.views.utils import ApiResponse
+from substrapp.views.utils import CharInFilter
 from substrapp.views.utils import ChoiceInFilter
 from substrapp.views.utils import MatchFilter
 from substrapp.views.utils import PermissionMixin
@@ -154,6 +156,12 @@ class AlgoRepFilter(FilterSet):
     category = ChoiceInFilter(
         field_name="category",
         choices=AlgoRep.Category.choices,
+    )
+
+    compute_plan_key = CharInFilter(field_name="compute_tasks__compute_plan__key", label="compute_plan_key")
+    dataset_key = CharFilter(field_name="compute_tasks__data_manager__key", distinct=True, label="dataset_key")
+    data_sample_key = CharInFilter(
+        field_name="compute_tasks__data_samples__key", distinct=True, label="data_sample_key"
     )
 
     class Meta: