+++ backend/substrapp/views/datamanager.py
@@ -3,6 +3,7 @@ from django.conf import settings
 from django.db import models
 from django.urls import reverse
 from django_filters.rest_framework import BaseInFilter
+from django_filters.rest_framework import CharFilter
 from django_filters.rest_framework import DateTimeFromToRangeFilter
 from django_filters.rest_framework import DjangoFilterBackend
 from django_filters.rest_framework import FilterSet
@@ -25,6 +26,7 @@ from substrapp.views.filters_utils import CustomSearchFilter
 from substrapp.views.filters_utils import LogsPermissionFilter
 from substrapp.views.filters_utils import ProcessPermissionFilter
 from substrapp.views.utils import ApiResponse
+from substrapp.views.utils import CharInFilter
 from substrapp.views.utils import MatchFilter
 from substrapp.views.utils import PermissionMixin
 from substrapp.views.utils import ValidationExceptionError
@@ -130,6 +132,13 @@ def create(request, get_success_headers):
 
 class DataManagerRepFilter(FilterSet):
     creation_date = DateTimeFromToRangeFilter()
+    compute_plan_key = CharInFilter(
+        field_name="compute_tasks__compute_plan__key", distinct=True, label="compute_plan_key"
+    )
+    algo_key = CharFilter(field_name="compute_tasks__algo__key", distinct=True, label="algo_key")
+    data_sample_key = CharInFilter(
+        field_name="compute_tasks__data_samples__key", distinct=True, label="data_sample_key"
+    )
 
     class Meta:
         model = DataManagerRep